Ensure status descriptions are not over 140 characters
It may not be clear that the `HALT` file contents are meant to be
formatted like a commit or PR, with a reasonable-length title and
optional body below.

When folks write things as one long line, that is interpreted as title
and used as the status description, and we get:

    Setting halted status for PR #37776
    Error: HttpError: Validation Failed: {"resource":"Status","code":"custom","field":"description","message":"description is too long (maximum is 140 characters)"}

We can just truncate to avoid it.